+++ Assets/Sound Effects/Attacks.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eb9a94a99776d1042bf5224ded7670b3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
